Se reparo los crash en el editor, ahora se pueden visualizar los componentes de compuerta
Haciendo doble click sobre las compuertas, se despliega un listado de los botones y receptores que forman parte del editor y pueden formar parte de su logica.

+++ editor/cmake.sh
@@ -1,8 +1,37 @@
 #!/bin/sh
-mkdir build
-cd build/
-cmake ..
-make
-./portal_editor
-cd ..
-rm -r build/
+
+if [[ $1 == "-d" ]]; then
+    mkdir Debug
+    cd Debug
+    cmake -DCMAKE_BUILD_TYPE=Debug ..
+    make
+fi
+
+if [[ $1 == "-ce" ]]; then
+    mkdir build
+    cd build/
+    cmake ..
+    make
+    ./portal_editor
+    cd ..
+    rm -r build/
+fi
+
+if [[ $1 == "-c" ]]; then
+    mkdir build
+    cd build/
+    cmake ..
+    make
+    cd ..
+fi
+
+if [[ $1 == "-e" ]]; then
+    cd build/
+    ./portal_editor
+    cd ..
+fi
+
+
+if [[ $1 == "-clean" ]]; then
+    rm -r build/
+fi
